+++ src/ReadonlyArray.ts
@@ -24,7 +24,7 @@ import { isSome, none, Option, some } from './Option'
 import { fromCompare, getMonoid as getOrdMonoid, Ord, ordNumber } from './Ord'
 import { ReadonlyNonEmptyArray } from './ReadonlyNonEmptyArray'
 import { Show } from './Show'
-import { TraversableWithIndex1 } from './TraversableWithIndex'
+import { TraversableWithIndex1, PipeableTraverseWithIndex1 } from './TraversableWithIndex'
 import { Unfoldable1 } from './Unfoldable'
 import { Witherable1 } from './Witherable'
 import { Traversable1, PipeableTraverse1 } from './Traversable'
@@ -1559,30 +1559,6 @@ const traverse_ = <F>(
   return (ta, f) => traverseWithIndexF(ta, (_, a) => f(a))
 }
 
-/**
- * @since 2.6.3
- */
-export const traverse: PipeableTraverse1<URI> = <F>(
-  F: Applicative<F>
-): (<A, B>(f: (a: A) => HKT<F, B>) => (ta: ReadonlyArray<A>) => HKT<F, ReadonlyArray<B>>) => {
-  const traverseF = traverse_(F)
-  return (f) => (ta) => traverseF(ta, f)
-}
-
-/**
- * @since 2.6.3
- */
-export const sequence: Traversable1<URI>['sequence'] = <F>(F: Applicative<F>) => <A>(
-  ta: ReadonlyArray<HKT<F, A>>
-): HKT<F, ReadonlyArray<A>> => {
-  return reduce_(ta, F.of(zero_()), (fas, fa) =>
-    F.ap(
-      F.map(fas, (as) => (a: A) => snoc(as, a)),
-      fa
-    )
-  )
-}
-
 const traverseWithIndex_ = <F>(F: Applicative<F>) => <A, B>(
   ta: ReadonlyArray<A>,
   f: (i: number, a: A) => HKT<F, B>
@@ -1848,6 +1824,40 @@ export const reduceRightWithIndex: <A, B>(b: B, f: (i: number, a: A, b: B) => B)
 // -------------------------------------------------------------------------------------
 // instances
 // -------------------------------------------------------------------------------------
+
+/**
+ * @since 2.6.3
+ */
+export const traverse: PipeableTraverse1<URI> = <F>(
+  F: Applicative<F>
+): (<A, B>(f: (a: A) => HKT<F, B>) => (ta: ReadonlyArray<A>) => HKT<F, ReadonlyArray<B>>) => {
+  const traverseF = traverse_(F)
+  return (f) => (ta) => traverseF(ta, f)
+}
+
+/**
+ * @since 2.6.3
+ */
+export const sequence: Traversable1<URI>['sequence'] = <F>(F: Applicative<F>) => <A>(
+  ta: ReadonlyArray<HKT<F, A>>
+): HKT<F, ReadonlyArray<A>> => {
+  return reduce_(ta, F.of(zero_()), (fas, fa) =>
+    F.ap(
+      F.map(fas, (as) => (a: A) => snoc(as, a)),
+      fa
+    )
+  )
+}
+
+/**
+ * @since 2.6.3
+ */
+export const traverseWithIndex: PipeableTraverseWithIndex1<URI, number> = <F>(
+  F: Applicative<F>
+): (<A, B>(f: (i: number, a: A) => HKT<F, B>) => (ta: ReadonlyArray<A>) => HKT<F, ReadonlyArray<B>>) => {
+  const traverseWithIndexF = traverseWithIndex_(F)
+  return (f) => (ta) => traverseWithIndexF(ta, f)
+}
 
 /**
  * @category instances