+++ CHANGELOG.md
@@ -21,6 +21,7 @@ high state of flux, you're at risk of it changing without notice.
   - move `pipe` to `function` module (@gcanti)
   - export `sequence` from all modules which admit a `Traversable` instance (@gcanti)
   - export a pipeable `traverse` function from all modules which admit a `Traversable` instance (@gcanti)
+  - export a pipeable `traverseWithIndex` function from all modules which admit a `TraversableWithIndex` instance (@gcanti)
 
 # 2.6.2